Add a standard time display
