+++ packages/frontend/src/components/CountDownTimer.component.tsx
@@ -1,5 +1,5 @@
-import { Typography } from '@mui/material';
 import { FC, useState, useEffect } from 'react';
+import { TimeDisplay } from './TimeDisplay.component';
 
 export type CountDownState = 'running' | 'paused' | 'restart';
 
@@ -10,22 +10,10 @@ export interface CountDownTimerProps {
 
 export const CountDownTimer: FC<CountDownTimerProps> = ({ seconds, status }) => {
   const [timeRemaining, setTimeRemaining] = useState<number>(seconds);
-  const [timeString, setTimeString] = useState<string>('00:00');
-
-  const updateTimeString = (seconds: number) => {
-    const minutesRemaining = Math.floor(seconds / 60);
-    const secondsRemaining = seconds - minutesRemaining * 60;
-
-    const minutesString = minutesRemaining.toString().padStart(2, '0');
-    const secondsString = secondsRemaining.toString().padStart(2, '0');
-
-    setTimeString(`${minutesString}:${secondsString}`);
-  };
 
   useEffect(() => {
     if (status == 'restart') {
       setTimeRemaining(seconds);
-      updateTimeString(seconds);
       return;
     }
   }, [status]);
@@ -37,7 +25,6 @@ export const CountDownTimer: FC<CountDownTimerProps> = ({ seconds, status }) => 
 
     if (status == 'running') {
       const newRemaining = timeRemaining - 1;
-      updateTimeString(newRemaining);
       setTimeRemaining(newRemaining);
     }
   };
@@ -47,11 +34,7 @@ export const CountDownTimer: FC<CountDownTimerProps> = ({ seconds, status }) => 
     return () => clearInterval(timerId);
   });
 
-  useEffect(() => {
-    updateTimeString(seconds);
-  }, []);
-
-  return <Typography variant="body1">{timeString}</Typography>;
+  return <TimeDisplay seconds={timeRemaining} />;
 };
 
 export default CountDownTimer;
